libs: Fix the submodule for tinyusb
It was impossible to build the tree because the tinyusb submodule
that was previously registered had broken builds. Alter the upstream
URL to wbcbz7's branch and pin to the rp2040-epx-only branch.

Additionally, actually add the rp2040-psram and tusb_xinput
submodules. They had no prior commit IDs in the tree, but were
defined in .gitmodules.